Cast the number of points to an int in the comparison with distance (since distance could be negative) to fix build in pipeline

+++ applications/GeoMechanicsApplication/custom_utilities/geometry_utilities.cpp
@@ -66,7 +66,7 @@ void ReverseNodes(InputIt                               Begin,
         GeometryFamily == GeometryData::KratosGeometryFamily::Kratos_Linear ? Begin : Begin + 1;
 
     const auto number_of_corner_points = GetNumberOfCornerPoints(GeometryFamily);
-    KRATOS_ERROR_IF(number_of_corner_points > std::distance(Begin, End))
+    KRATOS_ERROR_IF(static_cast<int>(number_of_corner_points) > std::distance(Begin, End))
         << "Number of nodes for reversal is too small for the geometry family and order type "
            "specified.\n";
     auto end_of_corner_points = Begin + number_of_corner_points;
@@ -77,7 +77,7 @@ void ReverseNodes(InputIt                               Begin,
         // For non-line geometries, there could be internal points as well, so we only reverse the
         // edge points here. For line geometries, all remaining points will be edge points.
         const auto number_of_edge_points = GetNumberOfEdgePoints(GeometryFamily, GeometryOrderType);
-        KRATOS_ERROR_IF(number_of_edge_points > std::distance(end_of_corner_points, End))
+        KRATOS_ERROR_IF(static_cast<int>(number_of_edge_points) > std::distance(end_of_corner_points, End))
             << "Number of nodes for reversal is too small for the geometry family and order type "
                "specified.\n";
         end_of_edge_points = end_of_corner_points + number_of_edge_points;
